test(native-trading): co-locate flow hooks
