+++ suite-native/module-trading/src/hooks/buy/useBuyInputFormControls.test.tsx
@@ -2,9 +2,9 @@ import { Form } from '@suite-native/forms';
 import { act, renderHookWithBasicProvider } from '@suite-native/test-utils';
 import { type BuyFormType } from '@suite-native/trading-types';
 
-import { renderHookWithTradingProvider } from '../../../__tests__/tradingTestUtils';
-import { useBuyForm } from '../useBuyForm';
-import { useBuyInputFormControls } from '../useBuyInputFormControls';
+import { useBuyForm } from './useBuyForm';
+import { useBuyInputFormControls } from './useBuyInputFormControls';
+import { renderHookWithTradingProvider } from '../../__tests__/tradingTestUtils';
 
 describe('useBuyInputFormControls', () => {
     let form: BuyFormType;
