+++ suite-native/module-trading/src/hooks/sell/useSellForm.test.tsx
@@ -23,8 +23,8 @@ import { selectTradingResidenceCountry } from '@suite-native/trading-state';
 import { type SellFormType } from '@suite-native/trading-types';
 import { PROTO } from '@trezor/connect';
 
-import { createTradingLightStore } from '../../../__tests__/tradingTestUtils';
-import { useSellForm } from '../useSellForm';
+import { useSellForm } from './useSellForm';
+import { createTradingLightStore } from '../../__tests__/tradingTestUtils';
 
 const btc1Account = getBtcAccount({ descriptor: asAccountDescriptor('btc1normal') });
 const eth1Account = getEthAccount({ descriptor: asAccountDescriptor('eth1normal') });
