+++ suite-native/module-trading/src/hooks/sell/useSellQuotes.test.ts
@@ -15,9 +15,9 @@ import {
 } from '@suite-native/trading-fixtures';
 import { type SellFormValues } from '@suite-native/trading-types';
 
-import { createTradingLightStore } from '../../../__tests__/tradingTestUtils';
-import { useSellForm } from '../useSellForm';
-import { useSellQuotes } from '../useSellQuotes';
+import { useSellForm } from './useSellForm';
+import { useSellQuotes } from './useSellQuotes';
+import { createTradingLightStore } from '../../__tests__/tradingTestUtils';
 
 const mockDebounce = (fn: () => unknown) => fn();
